requirements.txt.txt
flask
flask-mysqldb
pycryptodome
bcrypt